refactor account
