+++ rqalpha/mod/rqalpha_mod_sys_simulation/utils.py
@@ -16,11 +16,9 @@
 import datetime
 import six
 
-from rqalpha.const import DEFAULT_ACCOUNT_TYPE
 from rqalpha.model.positions import Positions
 from rqalpha.model.portfolio import Portfolio
 from rqalpha.model.trade import Trade
-from rqalpha.model.base_account import BaseAccount
 from rqalpha.utils.i18n import gettext as _
 
 from rqalpha.const import SIDE, POSITION_EFFECT
